+++ Makefile
@@ -15,7 +15,8 @@ SRCS = ./srcs/main.c  \
 		./srcs/error/error.c \
 		./srcs/game/init.c \
 		./srcs/game/render.c \
-		./srcs/game/mlx_utils.c
+		./srcs/utils/mlx_utils.c \
+		./srcs/player/player_move.c;
 
 OBJS = $(patsubst %.c, %.o, $(SRCS))
 INCDIRS = ./includes
